requirements.txt
palettable==3.1.1
isodate==0.6.0
pyspotlight==0.7.2
sphinx==1.7.5
sphinx_rtd_theme
docutils
regex==2018.2.21
SPARQLWrapper==1.8.2
pyyaml>=4.2b1
requests>=2.20.0
eliot
eliot-tree
